scheduler/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
asyncpg
alembic
greenlet
kafka-python
requests